-refactored scripts
